barra de pesquisa previamente funcionando, banco de dados completo

+++ index.js
@@ -1,3 +1,19 @@
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+//Green Life
+
+//Site original desenvolvido para apresentação da Atividade Prática Supervisionada (APS) da universidade Unip Alphaville, localizada em São Paulo, no primeiro semestre do curso de Sistemas de Informação com orientação da professora Sandra e professor Henry, na matéria de Programação Web Responsiva.
+
+//Não nos responsabilizamos por qualquer uso indevido de trechos de código retirados do projeto.
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+
+
+
+
+//REQUERIMENTO E CONFIGURAÇÃO DAS BIBLIOTECAS E MODULOS UTILIZADOS NO BACKEND
+
 var express = require("express");
 var app = express();
 
@@ -13,6 +29,12 @@ app.use(express.urlencoded());
 app.use(express.json());
 
 var multer = require("multer");
+
+
+
+
+
+//FUNÇÃO MULTER PARA MOVER O ARQUIVO RECEBIDO PARA A PASTA DESTINADA
 
 const storage = multer.diskStorage({
 	destination: function (req, file, cb) {
@@ -32,25 +54,30 @@ const storage = multer.diskStorage({
 	}
 });
 
+
+
+
+//VARIÁVEIS GLOBAIS UTILIZADAS NO BACKEND
+
 const upload = multer({ storage });
 
-function session(id){
-    axios.get(`http://localhost:7000/usuarios/?id=${id}`).then(resp => {
-
-    });
-}
-
-
-var log = false;
 var nome = "";
 var email = "";
 var senha = "";
 var image = "ftPerfil.png";
 
+
+
+
+
+//RENDERIZAÇÃO DA PAGINA INDEX
+
 app.get("/", function(req, res){
     var token = req.query.token;
     if(token === undefined){
         res.render("index.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token});
+    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+        res.render("index.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token});
     } else {
         axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
             var dados = resp.data[0];
@@ -64,6 +91,13 @@ app.get("/", function(req, res){
     }
     
 });
+
+
+
+
+
+
+//LOGIN/CADASTRO E PROCESSAMENTOS RELACIONADOS AO PERFIL
 
 app.post("/newUser", function(req, res){
     var nwnome = req.body.nome;
@@ -136,7 +170,9 @@ app.get("/adm", function(req, res){
 
 app.get("/perfil", function(req, res){
     var token = req.query.token;
-    if(token != undefined){
+    if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+        res.redirect("/adm/?token=e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248");
+    } else if(token != undefined){
         axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
             var dados = resp.data[0];
             console.log("Perfil: ", dados);
@@ -149,10 +185,6 @@ app.get("/perfil", function(req, res){
     } else {
         res.send("<html><head></head><body><script>alert(\"Faça login\");window.location.href='/'</script></body></html>");
     }
-});
-
-app.get("/log", function(req, res){
-    res.send(log);
 });
 
 app.get("/deslog", function(req, res){
@@ -194,12 +226,16 @@ app.post("/delet", function(req, res){
             var id = resp.data[0].id;
             var imagem = resp.data[0].image;
 
-            fs.unlink(`./public/fts/${imagem}`, (err => {
-                if (err) console.log(err);
-                else {
-                console.log(`\nDeleted file: ${imagem}`);
-                }
-            }));
+            if(imagem === "ftPerfil.png"){
+
+            } else {
+                fs.unlink(`./public/fts/${imagem}`, (err => {
+                    if (err) console.log(err);
+                    else {
+                    console.log(`\nDeleted file: ${imagem}`);
+                    }
+                }));
+            }
 
             axios.delete(`http://localhost:7000/usuarios/${id}`);
             nome = "";
@@ -236,10 +272,18 @@ app.post("/update", upload.single('ft'), function(req, res){
     res.send(`<html><head></head><body><script>alert("Alterado");window.location.href="/perfil/?token=${req.body.token}"</script></body></html>`);
 });
 
+
+
+
+
+//ROTEAMENTO PARA PAGINAS DO SITE
+
 app.get("/vidaSustentavel", function(req, res){
     var token = req.query.token;
     if(token === undefined){
         res.render("vidaSustentavel.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token});
+    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+        res.render("vidaSustentavel.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token});
     } else {
         axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
             var dados = resp.data[0];
@@ -257,6 +301,8 @@ app.get("/impactos", function(req, res){
     var token = req.query.token;
     if(token === undefined){
         res.render("impactos.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token});
+    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+        res.render("impactos.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token});
     } else {
         axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
             var dados = resp.data[0];
@@ -274,6 +320,8 @@ app.get("/sustentabilidade", function(req, res){
     var token = req.query.token;
     if(token === undefined){
         res.render("sustentabilidade.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token});
+    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+        res.render("sustentabilidade.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token});
     } else {
         axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
             var dados = resp.data[0];
@@ -291,6 +339,8 @@ app.get("/ds", function(req, res){
     var token = req.query.token;
     if(token === undefined){
         res.render("dsSustentavel.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token});
+    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+        res.render("dsSustentavel.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token});
     } else {
         axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
             var dados = resp.data[0];
@@ -308,6 +358,8 @@ app.get("/TIverde", function(req, res){
     var token = req.query.token;
     if(token === undefined){
         res.render("TIverde.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token});
+    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+        res.render("TIverde.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token});
     } else {
         axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
             var dados = resp.data[0];
@@ -325,6 +377,8 @@ app.get("/qs", function(req, res){
     var token = req.query.token;
     if(token === undefined){
         res.render("quemSomos.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token});
+    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+        res.render("quemSomos.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token});
     } else {
         axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
             var dados = resp.data[0];
@@ -337,5 +391,136 @@ app.get("/qs", function(req, res){
         })
     }
 });
+
+
+
+//BARRA DE PESQUISA
+
+
+app.get("/query", function(req, res){
+    var q = req.query.q;
+    var token = req.query.token;
+    axios.get(`http://localhost:5000/pesquisas?titulo_like=${q}`).then(resp_tit => {
+        
+        if(resp_tit.data[0] === undefined){
+
+            axios.get(`http://localhost:5000/pesquisas?conteudo_like=${q}`).then(resp_con => {
+
+                if(resp_con.data[0] === undefined){
+
+                    res.send(`<html><head></head><body><script>alert("Não há correspondências para sua pesquisa");window.location.href="/perfil/?token=${req.query.token}</script></body></html>`);
+                } else {
+
+                    for(let i = 0; i < resp_con.data.length; i++){
+                        resp_con.data[i].conteudo = resp_con.data[i].conteudo.substring(0, 200).concat('...');
+                    }
+
+                    if(token === undefined){
+
+                        res.render("querys.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token, "dados": resp_con.data});
+                    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+
+                        res.render("querys.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token, "dados": resp_con.data});
+                    } else {
+                        
+                        axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
+                            var dados = resp.data[0];
+                            console.log("2: ", dados);
+                            if(dados != undefined){
+                                res.render("querys.ejs", {"nome": dados.nome, "email": dados.email, "senha": dados.senha, "image": dados.image, "token": token, "dados": resp_con.data});
+                            } else {
+                                res.render("querys.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token, "dados": resp_con.data});
+                            }
+                        })
+                    }
+                }
+            
+            });
+
+        } else {
+
+            console.log(resp_tit.data);
+
+            axios.get(`http://localhost:5000/pesquisas?conteudo_like=${q}`).then(resp_con => {
+
+                if(resp_con.data[0] === undefined){
+
+                    for(let i = 0; i < resp_tit.data.length; i++){
+                        resp_tit.data[i].conteudo = resp_tit.data[i].conteudo.substring(0, 200).concat('...');
+                    }
+
+                    if(token === undefined){
+
+                        res.render("querys.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token, "dados": resp_tit.data});
+                    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+
+                        res.render("querys.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token, "dados": resp_tit.data});
+                    } else {
+                        
+                        axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
+                            var dados = resp.data[0];
+                            console.log("2: ", dados);
+                            if(dados != undefined){
+                                res.render("querys.ejs", {"nome": dados.nome, "email": dados.email, "senha": dados.senha, "image": dados.image, "token": token, "dados": resp_tit.data});
+                            } else {
+                                res.render("querys.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token, "dados": resp_tit.data});
+                            }
+                        })
+                    }
+                } else {
+
+                    titulo = resp_tit.data;
+                    conteudo = resp_con.data;
+
+                    for(let i = 0; i < titulo.length; i++){
+                        conteudo.push(titulo[i]);
+                    }
+                    var final = [];
+
+                    conteudo.forEach((item) => {
+                        var duplicado  = final.findIndex(redItem => {
+                            return item.id == redItem.id;
+                        }) > -1;
+
+                        if(!duplicado) {
+                            final.push(item);
+                        }
+                    });
+
+                    for(let i = 0; i < final.length; i++){
+                        final[i].conteudo = final[i].conteudo.substring(0, 200).concat('...');
+                    }
+
+                    if(token === undefined){
+
+                        res.render("querys.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token, "dados": final});
+                    } else if(token === "e1aa28844ce0b036eb2a31htrgrrgroglmvbrçlgmǵma02cb2223c34048b2b48ac20480a253f82167aa3d31142cbb9932a09dd5248"){
+
+                        res.render("querys.ejs", {"nome": "adm", "email": "adm@root.com", "senha": "root@root", "image": image, "token": token, "dados": final});
+                    } else {
+                        
+                        axios.get(`http://localhost:7000/usuarios/?token=${token}`).then(resp => {
+                            var dados = resp.data[0];
+                            console.log("2: ", dados);
+                            if(dados != undefined){
+                                res.render("querys.ejs", {"nome": dados.nome, "email": dados.email, "senha": dados.senha, "image": dados.image, "token": token, "dados": final});
+                            } else {
+                                res.render("querys.ejs", {"nome": nome, "email": email, "senha": senha, "image": image, "token": token, "dados": final});
+                            }
+                        })
+                    }
+                }
+            
+            });
+        }
+        
+    });
+});
+
+
+
+
+
+//CONFIGURAÇÃO DA PORTA EXPRESS A SER MONITORADA
 
 app.listen(3000);
